% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-borderpath.R
\name{draw_key_borderpath}
\alias{draw_key_borderpath}
\title{Key glyphs for legends}
\usage{
draw_key_borderpath(data, params, size)
}
\arguments{
\item{data, params, size}{See \code{ggplot2::draw_key_path()} for usage}
}
\description{
Key glyphs for legends
}
