% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-borderpath.R, R/ggproto-classes.R
\docType{data}
\name{GeomBorderpath}
\alias{GeomBorderpath}
\alias{GeomBorderline}
\alias{GeomBorderstep}
\alias{ggborderline-extensions}
\title{ggborderlines extensions to ggplot2}
\format{
An object of class \code{GeomBorderpath} (inherits from \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomBorderline} (inherits from \code{GeomBorderpath}, \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomBorderstep} (inherits from \code{GeomBorderpath}, \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomBorderpath

GeomBorderline

GeomBorderstep
}
\description{
ggborderlines makes use of the ggproto class system to extend the
functionality of ggplot2. In general the actual classes should be of little
interest to users as the standard ggplot2 api of using geom_* and stat_*
functions for building up the plot is encouraged.
}
\keyword{datasets}
