% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggborderline-package.R
\docType{package}
\name{ggborderline-package}
\alias{ggborderline}
\alias{ggborderline-package}
\title{ggborderline: Line Plots that Pop}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of geometries to make line plots a little bit nicer. Use along with 'ggplot2' to: - Improve the clarity of line plots with many overlapping lines - Draw more realistic worms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wurli/ggborderline}
  \item \url{https://wurli.github.io/ggborderline/}
}

}
\keyword{internal}
