% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-borderpath.R
\name{scale_border_colour_continuous}
\alias{scale_border_colour_continuous}
\alias{scale_border_colour_discrete}
\alias{scale_border_size_continuous}
\alias{scale_border_size_discrete}
\title{Scales for borderlines}
\usage{
scale_border_colour_continuous(...)

scale_border_colour_discrete(...)

scale_border_size_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 6),
  trans = "identity",
  guide = "legend"
)

scale_border_size_discrete(...)
}
\arguments{
\item{...}{Passed to the relevent ggplot2 scale}
}
\description{
These scales control the size and colour of the borders in borderlines. They
work in much the same way as \code{ggplot2::scale_colour_continuous()},
\code{ggplot2::scale_size_discrete()}, etc.
}
