% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-borderpath.R
\name{scale_bordercolour_continuous}
\alias{scale_bordercolour_continuous}
\alias{scale_bordercolour_discrete}
\alias{scale_borderwidth_continuous}
\alias{scale_borderwidth_discrete}
\title{Scales for borderlines}
\usage{
scale_bordercolour_continuous(..., aesthetics = "bordercolour")

scale_bordercolour_discrete(..., aesthetics = "bordercolour")

scale_borderwidth_continuous(..., aesthetics = "borderwidth")

scale_borderwidth_discrete(..., aesthetics = "borderwidth")
}
\arguments{
\item{...}{Passed to the corresponding ggplot2 scales}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful,
for example, to apply colour settings to the bordercolour and colour
aesthetics at the same time, via aesthetics = c("bordercolour", "colour").}
}
\value{
A ggproto scale object
}
\description{
These scales control the linewidth and colour of the borders in borderlines.
They work in much the same way as \code{ggplot2::scale_colour_continuous()},
\code{ggplot2::scale_linewidth_discrete()}, etc.
}
