% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-borderpath.R
\name{scale_border_colour_continuous}
\alias{scale_border_colour_continuous}
\alias{scale_border_colour_discrete}
\alias{scale_border_size_continuous}
\alias{scale_border_size_discrete}
\title{Scales for borderlines}
\usage{
scale_border_colour_continuous(..., aesthetics = "border_colour")

scale_border_colour_discrete(..., aesthetics = "border_colour")

scale_border_size_continuous(..., aesthetics = "border_size")

scale_border_size_discrete(..., aesthetics = "border_size")
}
\arguments{
\item{...}{Passed to the corresponding ggplot2 scales}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful,
for example, to apply colour settings to the border_colour and colour
aesthetics at the same time, via aesthetics = c("border_colour", "colour").}
}
\description{
These scales control the size and colour of the borders in borderlines. They
work in much the same way as \code{ggplot2::scale_colour_continuous()},
\code{ggplot2::scale_size_discrete()}, etc.
}
